bitthum/pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "study2-cloud-security-track"
version = "0.1.0"
description = "AWS-first cloud security junior track with learning labs and a control plane capstone"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
  "duckdb>=1.3.2",
  "fastapi>=0.116.1",
  "httpx>=0.28.1",
  "pydantic>=2.11.7",
  "psycopg[binary]>=3.2.9",
  "PyYAML>=6.0.2",
  "SQLAlchemy>=2.0.42",
  "typer>=0.16.1",
  "uvicorn>=0.35.0",
]

[project.optional-dependencies]
dev = [
  "mypy>=1.17.1",
  "pytest>=8.4.1",
  "ruff>=0.12.10",
]

[project.scripts]
study2-meta = "study2_meta:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
addopts = "-q"

[tool.ruff]
line-length = 100
target-version = "py313"

[tool.mypy]
python_version = "3.13"
strict = true
ignore_missing_imports = true

